+++ R/afl2.r
@@ -7,14 +7,15 @@
 #'
 #' @param variantlist Choose pre-defined list. reduced = ~20k SNPs that are 
 #' common in all super populations (default). hapmap3 = ~1.3 million hm3 SNPs
+#' @param opengwas_jwt Used to authenticate protected endpoints. Login to https://api.opengwas.io to obtain a jwt. Provide the jwt string here, or store in .Renviron under the keyname OPENGWAS_JWT.
 #'
 #' @export
 #' @return Data frame containing ancestry specific LD scores and allele frequencies for each variant
-afl2_list <- function(variantlist=c("reduced", "hapmap3")[1])
+afl2_list <- function(variantlist=c("reduced", "hapmap3")[1], opengwas_jwt=get_opengwas_jwt())
 {
 	if(variantlist == "reduced")
 	{
-		api_query("variants/afl2/snplist") %>% 
+		api_query("variants/afl2/snplist", opengwas_jwt=opengwas_jwt) %>% 
 			get_query_content() %>%
 			dplyr::as_tibble() %>%
 			return()
@@ -33,12 +34,13 @@ afl2_list <- function(variantlist=c("reduced", "hapmap3")[1])
 #'
 #' @param rsid Vector of rsids
 #' @param reference Default=`"1000g"`
+#' @param opengwas_jwt Used to authenticate protected endpoints. Login to https://api.opengwas.io to obtain a jwt. Provide the jwt string here, or store in .Renviron under the keyname OPENGWAS_JWT.
 #'
 #' @export
 #' @return Data frame containing ancestry specific LD scores and allele frequencies for each variant
-afl2_rsid <- function(rsid, reference="1000g")
+afl2_rsid <- function(rsid, reference="1000g", opengwas_jwt=get_opengwas_jwt())
 {
-	out <- api_query("variants/afl2", list(rsid=rsid)) %>% get_query_content()
+	out <- api_query("variants/afl2", list(rsid=rsid), opengwas_jwt=opengwas_jwt) %>% get_query_content()
 	if(inherits(out, "response"))
 	{
 		return(out)
@@ -54,12 +56,13 @@ afl2_rsid <- function(rsid, reference="1000g")
 #' @param chrpos list of `<chr>:<pos>` in build 37, e.g. `c("3:46414943", "3:122991235")`. 
 #' Also allows ranges e.g `"7:105561135-105563135"`
 #' @param reference Default=`"1000g"`
+#' @param opengwas_jwt Used to authenticate protected endpoints. Login to https://api.opengwas.io to obtain a jwt. Provide the jwt string here, or store in .Renviron under the keyname OPENGWAS_JWT.
 #'
 #' @export
 #' @return Data frame containing ancestry specific LD scores and allele frequencies for each variant
-afl2_chrpos <- function(chrpos, reference="1000g")
+afl2_chrpos <- function(chrpos, reference="1000g", opengwas_jwt=get_opengwas_jwt())
 {
-	out <- api_query("variants/afl2", list(chrpos=chrpos)) %>% get_query_content()
+	out <- api_query("variants/afl2", list(chrpos=chrpos), opengwas_jwt=opengwas_jwt) %>% get_query_content()
 	if(inherits(out, "response"))
 	{
 		return(out)
@@ -79,14 +82,15 @@ afl2_chrpos <- function(chrpos, reference="1000g")
 #' e.g. output from associations
 #' @param snpinfo Output from [`afl2_list`], [`afl2_rsid`] or [`afl2_chrpos`]. 
 #' If `NULL` then [`afl2_list()`] is used by default
+#' @param opengwas_jwt Used to authenticate protected endpoints. Login to https://api.opengwas.io to obtain a jwt. Provide the jwt string here, or store in .Renviron under the keyname OPENGWAS_JWT.
 #'
 #' @export
 #' @return data frame ordered by most likely ancestry, each row represents a super population and cor column represents the correlation between the GWAS dataset and the 1000 genomes super population allele frequencies
-infer_ancestry <- function(d, snpinfo=NULL)
+infer_ancestry <- function(d, snpinfo=NULL, opengwas_jwt=get_opengwas_jwt())
 {
 	if(is.null(snpinfo))
 	{
-		snpinfo <- afl2_list()
+		snpinfo <- afl2_list(opengwas_jwt=opengwas_jwt)
 	}
 	snpinfo <- snpinfo %>%
 		dplyr::inner_join(., d, by="rsid")
@@ -98,32 +102,5 @@ infer_ancestry <- function(d, snpinfo=NULL)
 	}) %>% sort(decreasing=TRUE)
 	out <- dplyr::tibble(pop=gsub("AF\\.", "", names(out)), cor=out)
 	return(out)
-}
-
-#' Look up sample sizes when meta data is missing from associations
-#'
-#' @param d Output from [`associations`]
-#'
-#' @export
-#' @return Data frame which is an updated version of the input but with sample sizes inferred where missing
-fill_n <- function(d)
-{
-	id <- d$id[1]
-	if(! "n" %in% names(d))
-	{
-		d$n <- NA
-	}
-	d$n <- as.numeric(d$n)
-	if(any(is.na(d$n)))
-	{
-		info <- gwasinfo(id)
-		if(!is.na(info$sample_size))
-		{
-			d$n <- info$sample_size
-		} else {
-			d$n <- info$ncase + info$ncontrol
-		}
-	}
-	return(d)	
 }