+++ NEWS.md
@@ -1,5 +1,6 @@
 # ieugwasr 1.0.0
 * Introducing JWT authorisation for the API
+* Phasing out Google Oauth2 authorisation
 * Added user() function to get user information
 * Fixing issue with anonymous functions and backwards compatibility
 * Bug in tophits when result is empty